requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
MouseInfo==0.1.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
opencv-python==4.6.0.66
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
Send2Trash==1.8.0
six==1.16.0
terminado==0.13.0
testpath==0.6.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.6.0
